iar: bug [XJ3-1272] fix iar enable interface can't make iar normal working caused by sif_mclk is disabled
Summary: optimize iar pixel clk config

Test Plan: /app/bin/demo

Reviewers: zhaojun.li

Reviewed By: zhaojun.li

Subscribers: rui.guo

Differential Revision: https://cr.hobot.cc/D46834

+++ drivers/misc/bt2hdmi/siHdmiTx_902x_TPI.h
@@ -1188,6 +1188,7 @@ int siHdmiTx_TPI_Init(void);
 void siHdmiTx_TPI_Poll(void);
 void siHdmiTx_VideoSel(byte vmode, byte VideoFormat);
 void siHdmiTx_AudioSel(byte Afs);
-int siHdmiTx_ReConfig(byte vmode, byte VideoFormat, byte Afs);
+int siHdmiTx_ReConfig(unsigned short vmode,
+		unsigned short VideoFormat, unsigned short Afs);
 void OnHdmiCableConnected(void);
 #endif
